Create event listeners for game board
Create event listeners for game board that load after the page
elements, attached to board IDs. Console logs which cell is clicked.

Create turn counter that starts at one, and increases with each click.

If turn counter is odd, the board cell's div HTML is replaced with X;
else, it is replaced with O.