Create update game logic
Create update game logic. Much of the current ugliness of the code
is due to troubleshooting there -- did not immediately realize that
the AJAX request needed a JSON clarification. Also did not realize that
there was some care needed regarding the update of the score; that is,
if I am using a stored value of the score to test for win functions,
if the browser handles that asynchronously, my win function was comparing
an outdated version of the game and not observing a win at expected
moves.

TLDR: I need to now update the win and ui files to run after game updates
are made, with the turn counter being passed appropriately.