Add UI turn indicator and change game navigation
Add a turn indicator that displays usernames (if sufficiently short)
and switches between players. Also change game navigation so that, when
midgame, your only option is to 'quit' the game, resetting the board.
This is the same reset event that occurs when you start a new game,
because the game is marked over independent of the reset (which just
affects the store.js file and the interface).